requirements.txt
aws-cdk.core==1.42.0
aws-cdk.aws-cognito==1.42.0
aws-cdk.aws-apigateway==1.42.0
aws-cdk.aws-lambda==1.42.0
aws-cdk.aws-iam==1.42.0
aws-cdk.aws-stepfunctions==1.42.0
aws-cdk.aws-stepfunctions-tasks==1.42.0
aws-cdk.aws-ses==1.42.0
aws-cdk.aws-ses-actions==1.42.0